requirements.txt
flask
wtforms
flask-wtf
flask-login
argon2-cffi
dnspython
gunicorn
requests
yarl
python-whois